fix: solve circular dependencies in cart services
-Create cart-validation service for commonly used validations
-Rename files to use singular form
-Remove default from interfaces
