+++ src/carts/types/cart-response.type.ts
@@ -1,8 +1,8 @@
 import { FormattedCartProduct } from "./formatted-cart-product.type";
 import { PriceChange } from "./price-change.type";
-import QuantityChange from "./quantity-change.type";
+import { QuantityChange } from "./quantity-change.type";
 
-export default interface CartResponse {
+export interface CartResponse {
     cartTotal: number;
     totalQuantity: number;
     products: FormattedCartProduct[];
