+++ src/common/errors/error-type.ts
@@ -2,6 +2,7 @@ export enum ErrorType {
     USER_NOT_FOUND = 'USER_NOT_FOUND',
     CART_NOT_FOUND = 'CART_NOT_FOUND',
     PRODUCT_NOT_FOUND = 'PRODUCT_NOT_FOUND',
+    CART_ITEM_NOT_FOUND = 'CART_ITEM_NOT_FOUND',
     OUT_OF_STOCK = 'OUT_OF_STOCK',
     QUANTITY_LIMIT_EXCEEDED = 'QUANTITY_LIMIT_EXCEEDED',
     STOCK_LIMIT_EXCEEDED = 'STOCK_LIMIT_EXCEEDED',
